Includes kitchen.php as well as a sample kitchen database
Created a sample DB in which orders are contained
these orders will be displayed on the kitchen table website
the main difference is including a time stamp on each item
so that kitchen staff can prioritise orders

-Design Choices to be made

--Delete irrelevant data from table? (e.g is price necessary when
showing orders to kitchen staff)
--Time format right now dates are in format such as 2100 (21:00 or 9PM)
Should this be changed?
Should the DB contain information such as order complete?
Or should completed orders simply be deleted from the kitchen page as they are completed?