Modal
In this sprint, I am a coauthor of queries for the refinements.  We did it together. And now my job is to do pop up part when the customers want know more about the dish. When he clicks on the dish name. There will pop up and show the ingredients and calories about the dish. But I haven't made it. The original wonder is setting a button beside the dish. In order to make the webpage clear. We decide to link the modal with the dish name. My next week job is to complete the pop up and do webpages.