Modifying PHP to get Image Paths from Database
The PHP code has been modified so that it selects a new ImagePath column from our menu table in our database.

This column stores the image path for each of our images. It means we can then use this image path in an img src tag to display the picture for each item that is being pulled from the database in the PHP.

Co-Authored-By: jeevanchakravarthy <jeevanchakravarthy@users.noreply.github.com>
Co-Authored-By: harshdeep <harshdeep3@users.noreply.github.com>
Co-Authored-By: gracebrumder <gracebrumder@users.noreply.github.com>
Co-Authored-By: hanjuwang <hanjuwang@users.noreply.github.com>
Co-Authored-By: rashid341 <rashid341@users.noreply.github.com>
Co-Authored-By: gbeckenstrater <gbeckenstrater@users.noreply.github.com>