Added StaffPages Folder (Contains Waiter Home Page)
Added StaffPages folder, which contains the Waiter home page (the page you get redirected to after a successful login from the Login page).

The folder also contains 4 blank PHP files which are webpages the staff member can be redirected to during usage of the waiter page. These 4 blank files are to be used for future PHP code as part of our Waiter view.