Create UI Document.pdf
This documents how the user can use the different interfaces in our website. The documentation contains screenshots of our product and highlights the actions the user can take to navigate through our site.And this week Jeevan Gordon and I finished the pop up part and croped the images in 300*300.

Co-Authored-By: gbeckenstrater <gbeckenstrater@users.noreply.github.com>
Co-Authored-By: jeevanchakravarthy <jeevanchakravarthy@users.noreply.github.com>
Co-Authored-By: alexmason98 <alexmason98@users.noreply.github.com>