Adding Popup Displaying Item Description, Ingredients, Allergens, and Calories
Adding a major commit which deals with multiple user stories/cards. This commit contains code which adds an Info button which when clicked displays a popup showing the item description, ingredients, allergens, and calories for each item.

The code is dynamic and not hard coded, because there is an SQL query which has a Cross Join on the menu item ID against the same ID in the descriptions, ingredients, allergens, and calories tables.

The CSS I wrote for the popup (containing the styling for the popup) has already been committed in pull request #43, which I have been co-authored in. The CSS being removed here is redundant as we use a different button now.