Updated Menu.txt
Some of the subcategories for the wines on our menu have been updated so that its easier to distinguish the difference types of wines we have when displaying the sub categories on the website.

Removed the accent from the 'e' in 'cafe' for item 299, as the website can misinterpret this character and display a question mark when printing the item's name.