Started Ordering System
This commit contains code for the ordering system on our website.

When the user clicks the 'Add to Cart' button on a particular item, it calls the addToCart div relating to the ID number. If the item is already in the basket (the TempOrders database table) then it doesn't insert the item into the database again and prints "This item is already in basket". If its not already in the user's basket, it adds the item to the basket by inserting it into the TempOrders database table.