Updated indexStylePage.css
I have added lots of CSS for the different popups on our website. Anything corresponding to '.overlay' is the styling for the popups on the index page where you view the items, and anything corresponding to '.basketOverlay' is the styling for the popups on the OrderPage.

I have also carried out styling improvements of already-existing divs/elements on the website to make the interface cleaner.