Updated Order Page to Include Ordering System
This update queries the database for any orders in the TempOrders table. If there are no items, it displays "No items in basket", otherwise it displays a list of the items.

I have also added code to handle the cancellation of an order. When cancel order button is pressed, a confirmation popup appears asking if the user is sure they want to cancel their order. Clicking Yes runs a "DELETE FROM" query and refreshes the page to display "No items in basket", whereas clicking No closes the popup and keeps displaying the order.