Implemented 'Pay After' Option on OrderPage.php
In this commit I have added a button which gives the customer the flexibility of paying for their order after they've eaten.

This code takes the items they've selected from the TempOrders table and inserts it into the final Orders table once they click "Pay After". There is error checking in place so if they try to place an order with no items in their basket it displays "There are no items in your order". I also check for key violation by checking if an ID in TempOrders exists in Orders from a recently placed order to not only prevent key violations but prevent two of the same item being added to a database table.