Updated ViewMenu.php
Ensured that the for loop creates as many Add to Cart divs as there are items. This is instead of using a magic number/hard coding the amount of times it should make an Add to Cart div.

I also added a timestamp so that we can see and make use of when the customer adds an item into their basket (which is stored in our TempOrders database table)