Added Table Entry Popup and Error Checking
Added table entry popups which asks the user for the table number when they click either the 'Pay Before' or 'Pay After' button, this will later be used to be inserted into the database.

I have added error checking to ensure the user enters a table number between the range of 1 - 10, which is how many tables we have. Also if they click the submit button without entering a valid table number or with no input, it will display "Please enter a valid table number". If the user enters 01 for example, there is a ltrim function to remove 0s to the left of the  other numbers so that we only insert 1 into the database.