Added Code which reads user's IP address to differentiate orders
This code means that on our payment system index page, we can get an order from our database based on the user's IP address. We then have an updated SQL query which selects all items where the stored IP is equal to the user's IP. This gets the items and prices ready to be processed by the payment system and bill the user for them.

Co-Authored-By: alexmason98 <alexmason98@users.noreply.github.com>