Integrated Databases Containing Orders with Payment System
This commit contains code which integrates the database tables storing our orders to our payment system.
On a successful payment, we insert the order from the TempOrders table into the Orders table (which is
where confirmed/successful orders go) based on the user's IP address, with the payment status set as 'Paid'.
The new order can then be displayed in the waiter and kitchen view.

Co-Authored-By: alexmason98 <alexmason98@users.noreply.github.com>