Bug Fixed Kitchen.php Page
There was a small bug where the Kitchen page sometimes stops displaying 2 rows instead of 1 when 1 of the rows is selected with the status 'Cooked'.

We think this bug may have occurred from either the SQL query or the HTTP refresh. The page now works correctly with the changes made.